chore: Clean up project page
