+++ assets/js/pages/ProjectPage/Navigation.tsx
@@ -5,7 +5,7 @@ import * as Icons from "@tabler/icons-react";
 import * as Spaces from "@/models/spaces";
 import { Paths } from "@/routes/paths";
 
-export default function Navigation({ space }: { space: Spaces.Space }) {
+export function Navigation({ space }: { space: Spaces.Space }) {
   return (
     <Paper.Navigation>
       <Paper.NavItem linkTo={Paths.spaceProjectsPath(space.id!)}>
